Mockito mocks using testable Deployment
Tests run directly in the container in this mode so beans can
always be replaced by mocks. However debugging such tests is
problematic as for some reason the debugger does not stop on
breakpoints. This is alsow mentioned in the README.md

(where tests run locally in the Arquillian container)